# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: AI Platform Builder API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /tasks:
    post:
      operationId: createTask
      tags: [tasks]
      summary: Create and run a task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTaskRequest"
      responses:
        "201":
          description: Created task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
  /tasks/{id}:
    get:
      operationId: getTask
      tags: [tasks]
      summary: Get task details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Task details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
    patch:
      operationId: updateTask
      tags: [tasks]
      summary: Update task title and model
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTaskRequest"
      responses:
        "200":
          description: Updated task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
    delete:
      operationId: deleteTask
      tags: [tasks]
      summary: Delete a task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /tasks/{id}/messages:
    get:
      operationId: getTaskMessages
      tags: [messages]
      summary: Get messages for a task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Messages
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageListResponse"
  /tasks/{id}/run:
    post:
      operationId: runTask
      tags: [tasks]
      summary: Execute a task with an agent via Ollama (SSE streaming)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RunTaskRequest"
      responses:
        "200":
          description: Streaming SSE response with thinking and output
          content:
            text/event-stream:
              schema:
                type: string
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success]
...
    Task:
      type: object
      properties:
        id:
          type: string
        projectId:
          type: string
        agentId:
          type: string
        title:
          type: string
        description:
          type: string
        status:
          type: string
        result:
          type: string
        model:
          type: string
        ollamaEndpoint:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, projectId, title, status, createdAt, updatedAt]
...
    CreateTaskRequest:
      type: object
      properties:
        projectId:
          type: string
        agentId:
          type: string
        title:
          type: string
        description:
          type: string
        model:
          type: string
        ollamaEndpoint:
          type: string
      required: [projectId, title]
    UpdateTaskRequest:
      type: object
      properties:
        title:
          type: string
        model:
          type: string
    RunTaskRequest:
      type: object
      properties:
        message:
          type: string
        model:
          type: string
        ollamaEndpoint:
          type: string
        enableThinking:
          type: boolean
      required: [message]
    Message:
      type: object
      properties:
        id:
          type: string
        taskId:
          type: string
        agentId:
          type: string
        role:
          type: string
        content:
          type: string
        createdAt:
          type: string
      required: [id, taskId, role, content, createdAt]
    MessageListResponse:
      type: object
      properties:
        messages:
          type: array
          items:
            $ref: "#/components/schemas/Message"
      required: [messages]
...