# lib/api-zod/src/generated/types/messageListResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Platform Builder API
 * OpenAPI spec version: 0.1.0
 */
import type { Message } from "./message";

export interface MessageListResponse {
  messages: Message[];
}